debugging tests
